cephadm: move get_container_info out of cephadm.py
Now that get_container_info is clean of dependencies on functions and
classes that only exist in cephadm.py we can move it.
I chose to move it to a new module, container_lookup.py, because I
didn't see any really good homes for it - other modules were too low
level or it didn't fit thematically.
